Make the Game resize the canvas to the necessary height itself
Previously, the GameView would resize the canvas and then the Game would resize itself to fit the new canvas dimensions.  This removes the Game's ability to adapt to any arbitrary canvas size, but now puts all the resizing in one place, and more importantly, ensures the canvas is the same size as the gameplay area.