Brighten blocks and background
This only brightens the sprite sheet, but the adjustment is also added to the tileable block PSD in case individual pieces need to be exported with the new brightness.  This seems like the best compromise between keeping the aesthetic and making the blocks more visible with screen brightness and contrast turned down.  Brighter seemed too bright at regular screen brightness.